requirements.txt
openai
python-telegram-bot==20.0